ecoride_flask/requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.2.1
Flask==3.1.1
Flask-Session==0.8.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgspec==0.19.0
psycopg==3.2.9
psycopg-pool==3.2.6
python-dotenv==1.1.1
typing_extensions==4.14.1
Werkzeug==3.1.3
